separate loss network and the main architecture
