+++ correction/main_correction.py
@@ -57,7 +57,7 @@ def run(cfg, dbinfo):
     if cfg['lTrain']:
         dHyper = {'batchSize': cfg['batchSize'], 'learningRate': cfg['lr'], 'epochs': cfg['epochs'],
                   'kl_weight': cfg['correction']['kl_weight'], 'perceptual_weight': cfg['correction']['perceptual_weight'],
-                  'pixel_weight': cfg['correction']['pixel_weight']}
+                  'pixel_weight': cfg['correction']['pixel_weight'], 'pl_network': cfg['correction']['pl_network']}
         if len(train_ref.shape) == 3:
             dData = {'train_ref': train_ref, 'test_ref': test_ref, 'train_art': train_art, 'test_art': test_art}
             cnn_main.fRunCNNCorrection(dData, sModelIn, patchSize, sOutPath, dHyper, cfg['lTrain'])
@@ -68,7 +68,7 @@ def run(cfg, dbinfo):
     else:
         dHyper = {'batchSize': cfg['batchSize'], 'bestModel': cfg['correction']['bestModel'],
                   'kl_weight': cfg['correction']['kl_weight'], 'perceptual_weight': cfg['correction']['perceptual_weight'],
-                  'pixel_weight': cfg['correction']['pixel_weight']}
+                  'pixel_weight': cfg['correction']['pixel_weight'], 'pl_network': cfg['correction']['pl_network']}
         test_ref = test_ref.reshape((-1, 1, patchSize[0], patchSize[1]))
         test_art = test_art.reshape((-1, 1, patchSize[0], patchSize[1]))
         dData = {'test_ref': test_ref, 'test_art': test_art}
