+++ utils/MotionCorrection/network.py
@@ -1,0 +1,20 @@
+from keras.layers import Conv2D, LeakyReLU, Conv2DTranspose
+
+def LeakyReluConv2D(filters, kernel_size, strides, padding):
+    def f(inputs):
+        conv2d = Conv2D(filters,
+                        kernel_size=kernel_size,
+                        strides=strides,
+                        padding=padding)(inputs)
+        return LeakyReLU()(conv2d)
+    return f
+
+def LeakyReluConv2DTranspose(filters, kernel_size, strides, padding):
+    def f(inputs):
+        conv2d = Conv2DTranspose(filters=filters,
+                                 kernel_size=kernel_size,
+                                 strides=strides,
+                                 padding=padding)(inputs)
+        return LeakyReLU()(conv2d)
+    return f
+
